requirements.txt
asgiref==3.4.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
